Did the last commit just fail to save? eitherway made a make blog post component and moved blogPosts to components need to make makeBlogPosts actually do what it says it does
