+++ src/components/makeBlogPosts.jsx
@@ -1,0 +1,9 @@
+
+
+// currently exports to blog
+export const MakePost = () => {
+    <div>
+        <input placeholder="Post Title"/>
+        <input placeholder="Post Content"/>
+    </div>
+}
